+++ src/pages/ui-components/List/index.mdx
@@ -5,7 +5,7 @@ name: List
 import InsertEmoticonIcon from '@mui/icons-material/InsertEmoticon';
 import { Playground } from 'dokz';
 
-import IconButton from '../ButtonIcon';
+import IconButton from '../IconButton';
 import List from './';
 
 # Lista
